migration functions (configure AMC13 and GLIB, start AMC13 and GLIB,sendBurstTriggers,Ntriggers) from scan routines to GEMTBUtils
