+++ gemdaq-testing/gemsupervisor/include/gem/supervisor/tbutils/GEMTBUtil.h
@@ -111,12 +111,23 @@ namespace gem {
 	  //SOAP MEssage AMC13	
 	  void sendInitializeMessageAMC13()
 	    throw (xgi::exception::Exception);
+	  void sendConfigureMessageAMC13()
+	    throw (xgi::exception::Exception);
+	  bool sendStartMessageAMC13()
+	    throw (xgi::exception::Exception);
+	  void sendAMC13trigger()
+	    throw (xgi::exception::Exception);
+	  void NTriggersAMC13()
+	    throw (xgi::exception::Exception);
 	  void sendStopMessageAMC13()
 	    throw (xgi::exception::Exception);
-	  
 
 	  //SOAP MEssage GLIB	
 	  void sendInitializeMessageGLIB()
+	    throw (xgi::exception::Exception);
+	  void sendConfigureMessageGLIB()
+	    throw (xgi::exception::Exception);
+	  bool sendStartMessageGLIB()
 	    throw (xgi::exception::Exception);
 	  void sendStopMessageGLIB()
 	    throw (xgi::exception::Exception);
