+++ gemdaq-testing/gemsupervisor/include/gem/supervisor/tbutils/LatencyScan.h
@@ -11,7 +11,6 @@ namespace gem {
 
       class LatencyScan : public GEMTBUtil
       {
-
 	//friend class GEMTBUtil
 	
       public:
@@ -20,23 +19,6 @@ namespace gem {
 	LatencyScan(xdaq::ApplicationStub * s)
 	  throw (xdaq::exception::Exception);
 	~LatencyScan();
-
-	//SOAP MEssage AMC13	
-	void sendConfigureMessageAMC13()
-	  throw (xgi::exception::Exception);
-	bool sendStartMessageAMC13()
-	  throw (xgi::exception::Exception);
-	void sendAMC13trigger()
-	  throw (xgi::exception::Exception);
-	void NTriggersAMC13()
-	  throw (xgi::exception::Exception);
-
-	//SOAP MEssage GLIB	
-	void sendConfigureMessageGLIB()
-	  throw (xgi::exception::Exception);
-	bool sendStartMessageGLIB()
-	  throw (xgi::exception::Exception);
-
 
 	// HyperDAQ interface
 	void webDefault(xgi::Input *in, xgi::Output *out)
