+++ gemdaq-testing/gemsupervisor/include/gem/supervisor/tbutils/ThresholdScan.h
@@ -21,24 +21,6 @@ namespace gem {
           throw (xdaq::exception::Exception);
         ~ThresholdScan();
 
-	//SOAP MEssage AMC13	
-	void sendConfigureMessageAMC13()
-	  throw (xgi::exception::Exception);
-	bool sendStartMessageAMC13()
-	  throw (xgi::exception::Exception);
-	void sendAMC13trigger()
-	  throw (xgi::exception::Exception);
-	void NTriggersAMC13()
-	  throw (xgi::exception::Exception);
-
-	//SOAP MEssage GLIB	
-	void sendConfigureMessageGLIB()
-	  throw (xgi::exception::Exception);
-	bool sendStartMessageGLIB()
-	  throw (xgi::exception::Exception);
-
-
-
         // HyperDAQ interface
         void webDefault(xgi::Input *in, xgi::Output *out)
           throw (xgi::exception::Exception);
